+++ packages/frontend/editor-ui/src/views/NodeView.vue
@@ -139,7 +139,7 @@ import CanvasChatButton from '@/components/canvas/elements/buttons/CanvasChatBut
 import { useFocusPanelStore } from '@/stores/focusPanel.store';
 import { useAITemplatesStarterCollectionStore } from '@/experiments/aiTemplatesStarterCollection/stores/aiTemplatesStarterCollection.store';
 import { useReadyToRunWorkflowsStore } from '@/experiments/readyToRunWorkflows/stores/readyToRunWorkflows.store';
-import { expressionsWorker } from '@/data/workers/instances';
+import { expressionsWorker } from '@/workers/instances';
 
 defineOptions({
 	name: 'NodeView',
@@ -339,8 +339,6 @@ async function initializeData() {
 
 	try {
 		await Promise.all(loadPromises);
-
-		expressionsWorker.run('setNodeTypes', JSON.stringify(nodeTypesStore.nodeTypes));
 	} catch (error) {
 		toast.showError(
 			error,
