+++ packages/frontend/editor-ui/src/workers/sqlite-wasm.d.ts
